+++ packages/create-dothtml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dothtml
 
+## 6.4.0
+
+### Minor Changes
+
+- Added theme builder pattern.
+
 ## 6.3.3
 
 ### Patch Changes